- Jorrit fixed several warnings in CEL caused by a few places 	  where size_t was not fixed. 	- Jorrit added a new iEntityLoader interface which is implemented 	  by the celentity addon. Using that interface one can now 	  more easilly access the celentity addon to load entities from 	  XML description files.

+++ cel/trunk/plugins/behaviourlayer/xml/xmlscript.cpp
@@ -1232,7 +1232,7 @@ bool celXmlScriptEventHandler::Execute (iCelEntity* entity,
   celBlXml* cbl = (celBlXml*)(behave->GetBehaviourLayer ());
   bool varprop_trace = cbl->varprop_trace;
 
-  int stack_size = stack.Length ();
+  size_t stack_size = stack.Length ();
   int i = startop;
   if (startop == 0)
   {
